Moved StringUtil to util package
