+++ backend/src/main/java/com/empire/Army.java
@@ -1,5 +1,6 @@
 package com.empire;
 
+import com.empire.util.StringUtil;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.ArrayList;
